Extracting Output component, adding missing fields to config forms.
